lint: the tree comes clean at /analyze and 183 clang-tidy checks
Nothing here changes what the tree does, except where it was doing the wrong
thing. MSVC's /analyze reports two findings over the whole tree and both are
in vendored noiseutils.cpp; clang-tidy at 183 of the 206 checks in bugprone,
performance, misc and readability reported roughly a thousand, and this is
what answering them came to.

Eleven were defects rather than habits.

image/reader/Bmp.cxx read the 8 bit palette into &colors - the address of the
local pointer - rather than into the buffer it pointed at, so a palettised
bmp wrote 1KB over the stack. The palette is a std::vector now, which also
frees it when a later read throws. No fixture is 8 bit, which is why it had
never been seen.

image/reader/Tga.cxx called memcmp for the second header signature and never
compared the result, so a file matching neither signature was accepted on the
strength of a non-zero return.

Image::Image computed width * height * bytesPerPixel in 32 bits before
assigning it to the uint64_t that holds it, and the jpeg reader indexed its
output through an unsigned int. Both are gone, along with the row offsets in
the png, tga and bmp paths.

Three event constructors were noexcept and allocate: Sound builds a
std::string from a string_view, and MouseButton and MouseMotion pass a string
literal to a base that takes std::string. A throw there was std::terminate,
not an exception. Sound.cpp also carried a #pragma once.

type/CameraProfile.cxx had no self-assignment guard on operator=.
font/Font2D.cxx widened a signed char to unsigned int, sign-extending \xFF to
0xFFFFFFFF before the special case caught it. voxel/src/Version.h defined
VOXEL_VERSION at namespace scope in a header, and voxel/src/main.cxx seeded
rand() from the clock, so two runs started in the same second played the same
world.

font/TextureFont.cxx leaked the freetype stroker on every failure between
FT_Stroker_New and the bitmap, and leaked the glyph FT_Get_Glyph built when
the atlas filled or a stroke failed. Both handles have an owner now -
StrokerHandle and GlyphHandle - so the release belongs to the scope rather
than to each of the five returns.

The three offline tools reported an uncaught exception as an abort with no
message in it. imagetool, moya and talyn each move their body into a run()
behind a try/catch, which reports through stdio: a last resort that can
itself throw is not one, and the iostreams can.

Nineteen functions scored above readability-function-cognitive-complexity's
threshold of 25 and are decomposed rather than suppressed.
RIBReader::request, at 110, becomes the eight groups the RI standard itself
uses, dispatched through a tri-state Result so that "not this group's" and
"failed" stay distinct. Project::read, ui::Engine::load, loadProperties and
resolveImages, BitmapFont::loadCharset, TextureFont::loadGlyphs,
Polygon::split, Bmp::read, PongScene::tick, PongEngine::handleEvent,
Engine::registerEventMappings, talyn's run and RotateManipulator::grab all
split along seams their own comments already named. Camera::updateVelocity
was the same block three times, once per axis, and is now one axisVelocity in
a loop; WingedEdgeBRep::addFace named the same two neighbours twice, once per
winding, and RenderContext::addPolygon ordered a bound's corners twice.

Six of the collapsed branch chains were saying something worth keeping, so
the collapse carries it as a comment. pong's setSingleplayerMode and
setMultiplayerMode have identical bodies because multiplayer is non-coop
today, and its four key binding branches are the unreachable input capture
items docs/TODO.md already records.

The rest is smaller: std::endl goes for a newline, with an explicit
std::flush kept on the three progress lines that precede a long render;
const RtToken becomes RtToken, matching both the RI standard and the sixty
other tokens in the same table that were never const; hand rolled clamps
become glm::min, glm::max, std::min and std::max; declarations lose the const
that only means something on a definition; else-after-return goes; and
membership tests become contains().

moya's RenderMan.cxx names out the parameters of ten RI stubs, which d0544fc
deliberately did not do - it took /wd4100 for the whole file instead, on the
grounds that a name commented out is undone as each function grows a body.
The target keeps /wd4100; clang-tidy does not read it, and these ten are the
ones it reports.

The gates that report all this arrive in the commit after this one. Against
the two together, a clean build of all 444 targets reports nothing at
/W4 /WX /analyze or at the 183 checks, ctest passes 22 of 22, and cpplint
is clean.

+++ api/image/Image.cxx
@@ -17,7 +17,7 @@ Image::Image() : data_(0), width_(0), height_(0), bpp_(24), format_(Format::RGB)
 
 Image::Image(uint32_t w, uint32_t h, uint8_t b) : width_(w), height_(h), bpp_(b) {
     unsigned int bytesPerPixel = bpp_ / 8;
-    uint64_t size = width_ * height_ * bytesPerPixel;
+    uint64_t size = static_cast<uint64_t>(width_) * height_ * bytesPerPixel;
     if (bytesPerPixel == 3) {
         format_ = Format::RGB;
     } else if (bytesPerPixel == 4) {
